[0.4.2] REFACTOR :: CHANGING THE SIGNATURE OF ALL VARIABLES AND REMOVING UNNECESSAY CODE
THE CHANGE WAS MADE IN THE CELLULAR SETTINGS AND IN THE OPERATING SYSTEM UPDATE.
IT CONSISTS OF SWITCHING THE HARDCODED SIGNATURE OF THE VARIABLES TO EXPORTED
VARIABLES THAT RECEIVE THE INSTANCE OF THE OBJECTS IN THE SCENE.OBJECTS THAT
WERE CREATED AND MODIFIED ONLY IN CODE NOW EXIST IN THE SCENE.

Body:
(*) phone_settings.gd -> all variables of scenes elements are exported and instantiated in the scene.
(*) phone_settings.tscn -> instantiation of exported variables
(*) update_os.gd -> all scene element variables are exported. Removal of code style settings.
(*) update_os.tscn -> instantiation of exported variables
(+) update_os.tscn -> addition of elements that were created in code, along with the configuration of the font

from: [0.4.1]